requirements.txt
torch>=1.10.0
numpy>=1.20.0
matplotlib>=3.4.0
kaggle-environments>=1.9.0
scikit-learn>=1.0.0
# chessnut>=0.3.1
python-chess>=1.9.0
seaborn